Large commit, implementations till ex 14

+++ ejb-module/src/main/java/com/realdolmen/ex/ejb/PassengerEJB.java
@@ -4,44 +4,76 @@ import com.realdolmen.ex.domain.Passenger;
 import com.realdolmen.ex.ejb.interfaces.RemotePassengerEJB;
 import com.realdolmen.ex.persistence.PassengerRepository;
 
-import javax.ejb.EJB;
 import javax.ejb.LocalBean;
 import javax.ejb.Stateless;
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
 import java.util.List;
 
 @Stateless
 @LocalBean
 public class PassengerEJB implements RemotePassengerEJB{
 
-    private PassengerRepository  passengerRepository;
+    @PersistenceContext
+    private EntityManager entityManager;
 
     @Override
     public List<Passenger> findPassengers() {
-        passengerRepository = new PassengerRepository();
-        return passengerRepository.findAll();
+        PassengerRepository passengerRepository = this.getRepo();
+        if(!passengerRepository.checkEntityManagerIsNull())
+            return passengerRepository.findAll();
+        else
+            return null;
     }
 
     @Override
     public Passenger findPassenger(Long id) {
-        passengerRepository = new PassengerRepository();
-        return passengerRepository.findById(id);
+        PassengerRepository passengerRepository = this.getRepo();
+        if(!passengerRepository.checkEntityManagerIsNull())
+            return passengerRepository.findById(id);
+        else
+            return null;
     }
 
     @Override
     public Passenger createPassenger(Passenger passenger) {
-        passengerRepository = new PassengerRepository();
-        return passengerRepository.create(passenger);
+        PassengerRepository passengerRepository = this.getRepo();
+        if(!passengerRepository.checkEntityManagerIsNull())
+            return passengerRepository.create(passenger);
+        else
+            return null;
     }
 
     @Override
     public void deletePassenger(Passenger passenger) {
-        passengerRepository = new PassengerRepository();
-        passengerRepository.delete(passenger);
+        PassengerRepository passengerRepository = this.getRepo();
+
+        if(!passengerRepository.checkEntityManagerIsNull())
+            passengerRepository.delete(passenger);
     }
 
     @Override
     public Passenger updatePassenger(Passenger passenger) {
-        passengerRepository = new PassengerRepository();
-        return passengerRepository.update(passenger);
+        PassengerRepository passengerRepository = this.getRepo();
+
+        if(!passengerRepository.checkEntityManagerIsNull())
+            return passengerRepository.update(passenger);
+        else
+            return null;
+    }
+
+    @Override
+    public void setEntityManager(EntityManager entityManager) {
+        this.entityManager = entityManager;
+    }
+
+    private PassengerRepository getRepo()
+    {
+        PassengerRepository passengerRepository = new PassengerRepository();
+        if(passengerRepository.checkEntityManagerIsNull())
+        {
+            passengerRepository.setEntityManager(entityManager);
+        }
+        return passengerRepository;
     }
 }
